// AngularMain/src/app/modules/models/LettersHdDto.ts
export interface LettersHdDto {
    tenentId: number;
    mytransid: number;
    letterType: number;
    senderReceiverParty: number;
    filledAt: number;
    locationId: number;
    employeeId: number;
    letterDated: Date;
    representative: string;
    receivedSentDate: Date;
    searchTag: string;
    description: string;
    dmsid: number;
    status: string;
    crupId: number;
    userid: string;
    active: boolean;
    entrydate: Date;
    entrytime: Date;
    updttime: Date;
    username: string;
}